refactor(security): Stage C.4 — split core/security_layer.py 23KB → package, all files ≤7.2KB
Splits the 23 KB monolithic core/security_layer.py into a 4-module
package along the policies / audit / detection / abac seam. Resolves
PLATFORM_READINESS Gate v0.3 dimension A for this file (4th of five
violations after C.1 wiki_generator, C.2 cascade, C.3 character).

Module-size gate now down to 1 remaining violation (graph_editor.py
at the 20 KB boundary).

## What changed

- core/security_layer.py (23 KB, 508 lines, 12 module-level functions +
  9 constants + 1 SecurityLayer class) → core/security_layer/ package
  preserving the entire external surface byte-identical.
- No function signatures, return shapes, attribute names, or
  behaviour changed. Pure refactor.

## File sizes (post-split)

  __init__.py     7.2 KB   facade + SecurityLayer class
  _abac.py        7.1 KB   ABAC + output filter + cross-stage verify
  _audit.py       1.6 KB   log_attack / log_system_event wrappers
  _detection.py   4.6 KB   validate / detect_attack / detect_risky /
                           extract_data_only / sanitize_document_content
  _policies.py    6.5 KB   ROLE/SENSITIVITY tables + ATTACK_PATTERNS +
                           ATTACK_REGEX + RISKY_CODING_REGEX +
                           INSTRUCTION_INJECTION_PATTERNS +
                           SENSITIVE_PATTERNS + role keyword tables

All five files comfortably under the 20 KB gate.

## External surface preserved

All ~20 import paths work byte-identical, including:

- from core.security_layer import SecurityLayer                    (15+ sites)
- from core.security_layer import log_attack, log_system_event     (audit-bridge consumers)
- from core.security_layer import detect_attack, check_access      (graph_rag_engine, policy_engine)
- from core.security_layer import detect_risky_coding              (tests/test_risky_coding_policy.py)
- from core.security_layer import cross_stage_abac_verify          (tests/test_multimodal_trusted.py)
- from core.security_layer import extract_data_only,
                                  sanitize_document_content        (core/policy_engine.py)
- from core.security_layer import ROLE_LEVEL                       (server_llmwiki.py)
- from core.security_layer import mask_sensitive                   (server_llmwiki.py)

The package __init__.py re-exports every public name (and the
underscore-free constants) so existing call-sites stay unchanged.

## Module layout (Layer → File)

- L0 constants (no deps)        → _policies.py
- L0 audit wrappers (audit_bridge only) → _audit.py
- L1 detection (uses _policies + _audit) → _detection.py
- L1 ABAC + output filter (uses _policies + _audit + lazy policy_engine) → _abac.py
- L2 SecurityLayer class (uses everything above) → __init__.py

No circular deps. lazy imports for ``core.policy_engine`` preserved
verbatim (policy_engine still imports detection.extract_data_only +
audit.log_attack from this module — both visible at the package
level).

## Verification

- Full pytest suite (CI ignore list): **1590 passed + 823 subtests
  passed**. Zero regression.
- scripts/dogfood_check.py — 4/4 plugin invariants still PASS.
- ruff check core/security_layer/ — clean.

## Stage C progress

- C.1 (wiki_generator.py 51 KB) — ✅ PR #427
- C.2 (cascade.py 24 KB) — ✅ PR #428
- C.3 (character_profile.py 24 KB) — ✅ PR #429
- C.4 (security_layer.py 23 KB) — ✅ this PR
- C.5 (graph_editor.py 20 KB boundary) — pending; reassess

**4 of 5 module-size violations cleared on origin/main.**

## Out of scope

- Behavioural changes to SecurityLayer pipeline — pure refactor.
- PolicyEngine lazy-import boundary — preserved verbatim.
